requirements.txt
allure-pytest==2.14.2
allure-python-commons==2.14.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
curlify==2.2.1
distconfig3==1.0.1
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pydantic==2.11.5
pydantic_core==2.33.2
pydentic==0.0.1.dev3
PyHamcrest==2.1.0
pytest==8.3.5
python-stdnum==2.1
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
six==1.17.0
structlog==25.3.0
toml==0.10.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
vyper-config==1.2.1
watchdog==6.0.0
